introduce `fold_goal` and `fold_program_clause` callbacks
